api/requirements.txt
# Core Framework
fastapi==0.104.1

# AI/ML
openai==1.3.7

# HTTP and API
httpx==0.25.2
requests==2.31.0
python-multipart==0.0.6

# Data handling and validation
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0
python-dotenv==1.0.0

# Date/Time handling
python-dateutil==2.8.2

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
pytz==2023.3

